lock dependency to ibmiotf 0.2.39
This version has been tested in the past, so we might want
to stick to it.
If we do an update now, we get 0.2.41 instead which doesn't even log
the disconnections on the client side.
It won't even detect the disconnection...
Otherwise we get a different result every time we do a new install.
See https://github.com/dmelpi/bm/issues/190.